pi-voice-assistant/requirements.txt
pvporcupine==3.0.2
pyaudio==0.2.14
numpy==1.24.3
websockets==12.0
requests==2.31.0
python-dotenv==1.0.0
